--- services/ui-src/src/components/reports/EntityOverlayPage.tsx
@@ -1,143 +1,0 @@
-import { useState } from "react";
-// components
-import {
-  Box,
-  Button,
-  Flex,
-  Image,
-  Spinner,
-  useDisclosure,
-} from "@chakra-ui/react";
-import { Alert, Form, ReportPageIntro, CloseEntityModal } from "components";
-// types
-import { EntityShape, AlertTypes, EntityOverlayPageShape } from "types";
-// assets
-import closeIcon from "assets/icons/icon_cancel_x_white.png";
-import arrowLeftBlue from "assets/icons/icon_arrow_left_blue.png";
-
-export const EntityOverlayPage = ({ route, validateOnRender }: Props) => {
-  const submitting = false;
-  const [selectedEntity, setSelectedEntity] = useState<EntityShape | undefined>(
-    undefined
-  );
-  const { form, verbiage } = route;
-
-  // add/edit entity modal disclosure and methods
-  const {
-    isOpen: closeEntityModalIsOpen,
-    onOpen: closeEntityModalOnOpenHandler,
-    onClose: closeEntityModalOnCloseHandler,
-  } = useDisclosure();
-
-  const openCloseEntityModal = (entity?: EntityShape) => {
-    setSelectedEntity(entity);
-    closeEntityModalOnOpenHandler();
-  };
-
-  const closeCloseEntityModal = () => {
-    setSelectedEntity(undefined);
-    closeEntityModalOnCloseHandler();
-  };
-
-  return (
-    <Box>
-      <Button
-        sx={sx.backButton}
-        variant="none"
-        //TO-DO: add onClick prop to go back to initiative dashboard
-        aria-label="Return to all initiatives"
-      >
-        <Image src={arrowLeftBlue} alt="Arrow left" sx={sx.backIcon} />
-        Return to all initiatives
-      </Button>
-
-      {verbiage.intro && <ReportPageIntro text={verbiage.intro} />}
-      <Form
-        id={form.id}
-        formJson={form}
-        onSubmit={() => {}}
-        autosave
-        validateOnRender={validateOnRender || false}
-        dontReset={false}
-      />
-      <Box>
-        {verbiage.closeOutWarning && (
-          <Alert
-            title={verbiage.closeOutWarning.title}
-            status={AlertTypes.WARNING}
-            description={verbiage.closeOutWarning.description}
-          />
-        )}
-      </Box>
-
-      <Box>
-        {verbiage.closeOutModal && (
-          <Box>
-            <Button
-              rightIcon={
-                <Image src={closeIcon} alt="Close" sx={sx.closeIcon} />
-              }
-              onClick={() => openCloseEntityModal()}
-            >
-              {verbiage.closeOutModal.closeOutModalButtonText}
-            </Button>
-
-            <CloseEntityModal
-              selectedEntity={selectedEntity}
-              verbiage={verbiage}
-              modalDisclosure={{
-                isOpen: closeEntityModalIsOpen,
-                onClose: closeCloseEntityModal,
-              }}
-            />
-          </Box>
-        )}
-      </Box>
-
-      <Box sx={sx.footerBox}>
-        <Flex sx={sx.buttonFlex}>
-          <Button type="submit" form={form.id} sx={sx.saveButton}>
-            {submitting ? <Spinner size="md" /> : "Save & return"}
-          </Button>
-        </Flex>
-      </Box>
-    </Box>
-  );
-};
-
-interface Props {
-  route: EntityOverlayPageShape;
-  validateOnRender?: boolean;
-}
-
-const sx = {
-  backButton: {
-    padding: 0,
-    fontWeight: "normal",
-    color: "palette.primary",
-    display: "flex",
-    position: "relative",
-    right: "3rem",
-    marginBottom: "2rem",
-    marginTop: "-2rem",
-  },
-  backIcon: {
-    color: "palette.primary",
-    height: "1rem",
-    marginRight: "0.5rem",
-  },
-  closeIcon: {
-    width: "0.85rem",
-  },
-  footerBox: {
-    marginTop: "2rem",
-    borderTop: "1.5px solid var(--chakra-colors-palette-gray_light)",
-  },
-  buttonFlex: {
-    justifyContent: "end",
-    marginY: "1.5rem",
-  },
-  saveButton: {
-    width: "8.25rem",
-  },
-};
