+++ services/ui-src/src/components/reports/ModalOverlayReportPage.tsx
@@ -5,7 +5,7 @@ import {
   AddEditEntityModal,
   Alert,
   DeleteEntityModal,
-  EntityDetailsOverlay,
+  EntityDetailsDashboardOverlay,
   EntityProvider,
   EntityRow,
   ReportPageFooter,
@@ -131,7 +131,7 @@ export const ModalOverlayReportPage = ({
     <Box>
       {overlayForm && isEntityDetailsOpen && currentEntity ? (
         <EntityProvider>
-          <EntityDetailsOverlay
+          <EntityDetailsDashboardOverlay
             closeEntityDetailsOverlay={closeEntityDetailsOverlay}
             entityType={entityType as EntityType}
             entities={report?.fieldData[entityType]}
